v0.4.9: filter 2d workflow: bug-fix

+++ squirrel/library/affine_matrices.py
@@ -679,7 +679,7 @@ class AffineMatrix:
         return self._ndim
 
     def inverse(self):
-        inv = np.linalg.inv(self.get_matrix(order='Ms').astype(np.float128))
+        inv = np.linalg.inv(self.get_matrix(order='Ms').astype(np.float64))
         inv = self._ms_to_c(inv)
         return AffineMatrix(parameters=inv, pivot=self.get_pivot())
